# Optimise `ConcurrentWeakHashSet`