Benchmarks run with:
```bash
sudo java -Djava.library.path=async-profiler/lib -XX:+UnlockDiagnosticVMOptions -XX:+DebugNonSafepoints -Xms2G -Xmx2G -XX:+AlwaysPreTouch -jar benchmarks/target/scala-2.13/benchmarks.jar zio.internal.ConcurrentWeakHashSet
```

Machine:
<img width="236" alt="Screenshot 2024-12-13 at 11 01 53 PM" src="https://github.com/user-attachments/assets/9435957b-d14b-4be7-80f0-91f09a6e2f16" />


Before:
```scala
Benchmark                                                         Mode  Cnt        Score       Error   Units
ConcurrentWeakHashSetAddBenchmark.javaAddConcurrent              thrpt   10     3228.906 ±   291.269  ops/ms
ConcurrentWeakHashSetAddBenchmark.javaAddSerial                  thrpt   10     5044.286 ±   963.319  ops/ms
ConcurrentWeakHashSetAddBenchmark.springAddConcurrent            thrpt   10     2347.606 ±   193.925  ops/ms
ConcurrentWeakHashSetAddBenchmark.springAddSerial                thrpt   10     2679.433 ±   365.380  ops/ms
ConcurrentWeakHashSetAddBenchmark.zioAddConcurrent               thrpt   10     3409.547 ±   325.511  ops/ms
ConcurrentWeakHashSetAddBenchmark.zioAddSerial                   thrpt   10     4116.362 ±    72.991  ops/ms
ConcurrentWeakHashSetContainsBenchmark.javaContainsConcurrent    thrpt   10    26202.002 ±  9431.657  ops/ms
ConcurrentWeakHashSetContainsBenchmark.javaContainsSerial        thrpt   10   199349.743 ±  4017.164  ops/ms
ConcurrentWeakHashSetContainsBenchmark.springContainsConcurrent  thrpt   10  1254915.047 ± 13394.385  ops/ms
ConcurrentWeakHashSetContainsBenchmark.springContainsSerial      thrpt   10   215392.210 ±   932.719  ops/ms
ConcurrentWeakHashSetContainsBenchmark.zioContainsConcurrent     thrpt   10  1424593.379 ± 19148.211  ops/ms
ConcurrentWeakHashSetContainsBenchmark.zioContainsSerial         thrpt   10   252539.622 ±  1182.431  ops/ms
ConcurrentWeakHashSetIterateBenchmark.javaIterateConcurrent      thrpt   10      293.601 ±    16.399  ops/ms
ConcurrentWeakHashSetIterateBenchmark.javaIterateSerial          thrpt   10      425.283 ±    20.492  ops/ms
ConcurrentWeakHashSetIterateBenchmark.springIterateConcurrent    thrpt   10      847.324 ±   107.718  ops/ms
ConcurrentWeakHashSetIterateBenchmark.springIterateSerial        thrpt   10      153.034 ±     3.785  ops/ms
ConcurrentWeakHashSetIterateBenchmark.zioIterateConcurrent       thrpt   10     1178.970 ±    21.310  ops/ms
ConcurrentWeakHashSetIterateBenchmark.zioIterateSerial           thrpt   10      207.289 ±     1.098  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.javaRemoveConcurrent        thrpt   10     7877.045 ±  1498.372  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.javaRemoveSerial            thrpt   10    34200.363 ±   712.461  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.springRemoveConcurrent      thrpt   10      173.868 ±     2.151  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.springRemoveSerial          thrpt   10       29.006 ±     1.714  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.zioRemoveConcurrent         thrpt   10    11944.236 ±   637.579  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.zioRemoveSerial             thrpt   10    13130.778 ±   211.017  ops/ms
```

After:
```scala
Benchmark                                                         Mode  Cnt        Score       Error   Units
ConcurrentWeakHashSetAddBenchmark.javaAddConcurrent              thrpt   10     3203.185 ±   180.129  ops/ms
ConcurrentWeakHashSetAddBenchmark.javaAddSerial                  thrpt   10     4411.808 ±   913.229  ops/ms
ConcurrentWeakHashSetAddBenchmark.springAddConcurrent            thrpt   10     2394.420 ±   144.192  ops/ms
ConcurrentWeakHashSetAddBenchmark.springAddSerial                thrpt   10     2631.986 ±   329.917  ops/ms
ConcurrentWeakHashSetAddBenchmark.zioAddConcurrent               thrpt   10     3303.541 ±   540.057  ops/ms
ConcurrentWeakHashSetAddBenchmark.zioAddSerial                   thrpt   10     3884.573 ±   211.591  ops/ms
ConcurrentWeakHashSetContainsBenchmark.javaContainsConcurrent    thrpt   10    25973.355 ± 10733.229  ops/ms
ConcurrentWeakHashSetContainsBenchmark.javaContainsSerial        thrpt   10   199715.706 ±  2889.299  ops/ms
ConcurrentWeakHashSetContainsBenchmark.springContainsConcurrent  thrpt   10  1252340.898 ± 22503.605  ops/ms
ConcurrentWeakHashSetContainsBenchmark.springContainsSerial      thrpt   10   221062.054 ±   983.634  ops/ms
ConcurrentWeakHashSetContainsBenchmark.zioContainsConcurrent     thrpt   10  1424895.595 ± 15642.364  ops/ms
ConcurrentWeakHashSetContainsBenchmark.zioContainsSerial         thrpt   10   251443.873 ±  1795.220  ops/ms
ConcurrentWeakHashSetIterateBenchmark.javaIterateConcurrent      thrpt   10      315.851 ±    69.380  ops/ms
ConcurrentWeakHashSetIterateBenchmark.javaIterateSerial          thrpt   10      432.246 ±    18.140  ops/ms
ConcurrentWeakHashSetIterateBenchmark.springIterateConcurrent    thrpt   10      839.883 ±    69.762  ops/ms
ConcurrentWeakHashSetIterateBenchmark.springIterateSerial        thrpt   10      159.452 ±     2.530  ops/ms
ConcurrentWeakHashSetIterateBenchmark.zioIterateConcurrent       thrpt   10     1182.894 ±    16.480  ops/ms
ConcurrentWeakHashSetIterateBenchmark.zioIterateSerial           thrpt   10      207.925 ±     3.254  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.javaRemoveConcurrent        thrpt   10     7987.385 ±  1138.683  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.javaRemoveSerial            thrpt   10    35202.307 ±   663.493  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.springRemoveConcurrent      thrpt   10      203.521 ±     2.944  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.springRemoveSerial          thrpt   10       27.215 ±     1.284  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.zioRemoveConcurrent         thrpt   10    14748.800 ±   282.336  ops/ms
ConcurrentWeakHashSetRemoveBenchmark.zioRemoveSerial             thrpt   10    21598.214 ±   329.238  ops/ms
```

Seems like only `zioRemoveConcurrent` and `zioRemoveSerial` are profiting from these changes