@HollandDM 2024-12-24
Sorry if this sounds dumb, but why doesn't ZIO use this for `FiberRuntime`'s children? AFAIK this isn't used any where, and we're having a bounty for something look just like this?