@kyri-petrou 2024-12-24
> Sorry if this sounds dumb, but why doesn't ZIO use this for `FiberRuntime`'s children? AFAIK this isn't used anywhere, and we're having a [bounty PR](https://github.com/zio/zio/issues/8861) for something look just like this?

@HollandDM yeah that's a good question that I wondered myself in the past about. From my point of view, the `ConcurrentWeakHashSet` is better than Java's for concurrent add / remove / iterate, but not as good when these methods are invoked by a single thread. In the case of `FiberRuntime`, modification and iteration of the set is performed by a single thread, so in theory Java's set should perform better