@hearnadam 2025-03-17
@guizmaii would you mind sharing the before/after jmh result json files? You can load them via gist here: https://jmh.morethan.io/